Specify a prefix for names defined by the package
By default, this is implicitly derived from the buffer being linted,
but as this would signal warnings for multi-file packages, there is
also already the ability to specify the "main" module for a package,
which is then used to derive the common prefix for all package
modules. Yet, this too is inadequate for component packages in a
package suite that all share a common prefix.

Explicitly specifying the prefix allows the linter to easily check
such cases.